RLM-316 Implement strict/loose artifact options
When deploying to a new environment it is often the case
that the base OS installed has packages which are newer
than those available in the apt artifacts.

To ease the initial installation and provide the ability
to transition to using the apt artifacts exclusively, an
option is made available to do a 'loose' mode for the apt
artifacts. This mode leaves the existing apt sources in
place, ensuring that we don't end up with apt returning
that packages are installed which are not available in
the sources configured as it does when in 'strict' mode.

The facility to switch between modes is also implemented
so that an initial implementation trying to use strict
mode can switch to loose mode and continue the deployment.

A little bit of tidying up to the AIO boostrap has been
included in order to make it easier to read.